pyproject.toml
[project]
name = "devflow"
version = "0.2.0"
description = "AI agent for streamlining software development workflows"
readme = "README.md"
authors = [
    {name = "João Ferreira", email = "joaosequeiraferreira@icloud.com"}
]
requires-python = ">=3.13,<4.0"
dependencies = [
    "google-adk>=1.19.0",
    "google-genai>=1.52.0",
    "litellm>=1.80.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
]

[tool.poetry.group.dev.dependencies]
google-adk = {version = ">=1.19.0", extras = ["eval"]}
pytest = ">=9.0.1"
pytest-asyncio = ">=1.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = "."
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
